🐛 fix: type 정리 #63
