+++ src/pages/admin/api/useAdminData.ts
@@ -1,10 +1,10 @@
-import { MainMatchType, PageInfoProps } from "@/common/types/matchTypes";
+import { AdminMatchType, PageInfoProps } from "@/common/types/matchTypes";
 import axiosInstance from "@/common/utils/axiosInstance";
 import { useSuspenseInfiniteQuery } from "@tanstack/react-query";
 
 export default function useAdminData(selectedSport: string = "All") {
   return useSuspenseInfiniteQuery<{
-    content: MainMatchType[];
+    content: AdminMatchType[];
     pageInfo: PageInfoProps;
   }>({
     queryKey: ["matches", selectedSport],
